+++ .prototools
@@ -1,3 +1,3 @@
 [plugins]
 moon-test = "source:https://raw.githubusercontent.com/moonrepo/moon/master/proto-plugin.toml"
-wasm-test = "source:./plugins/target/wasm32-wasi/debug/proto_wasm_test.wasm"
+# wasm-test = "source:./plugins/target/wasm32-wasi/debug/proto_wasm_test.wasm"
