+++ crates/cli/src/commands/add_plugin.rs
@@ -10,5 +10,5 @@ pub async fn add_plugin_old() {
         color::shell("proto tool add")
     );
 
-    tool::add_tool(states, resources, emitters).await?;
+    tool::tool_add(states, resources, emitters).await?;
 }
