+++ crates/cli/src/commands/tool/add.rs
@@ -20,7 +20,7 @@ pub struct AddToolArgs {
 }
 
 #[system]
-pub async fn add_tool(args: ArgsRef<AddToolArgs>) {
+pub async fn tool_add(args: ArgsRef<AddToolArgs>) {
     if args.global {
         let mut user_config = UserConfig::load()?;
         user_config
