+++ crates/cli/src/commands/tool/mod.rs
@@ -1,9 +1,11 @@
 mod add;
 mod info;
+mod list;
 mod plugins;
 mod remove;
 
 pub use add::*;
 pub use info::*;
+pub use list::*;
 pub use plugins::*;
 pub use remove::*;
